# services/engine/openapi.yaml
openapi: 3.1.0
info:
  title: NewGRIP Engine API
  version: 0.3.0
paths:
  /health:
    get:
      responses:
        "200":
          description: OK
  /v1/operations:
    get:
      summary: List available operations
      responses:
        "200":
          description: Operation catalog
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    op:
                      type: string
                    category:
                      type: string
                    inputs:
                      type: array
                      items:
                        $ref: "#/components/schemas/SocketSpec"
                    outputs:
                      type: array
                      items:
                        $ref: "#/components/schemas/SocketSpec"
                    params:
                      type: array
                      items:
                        $ref: "#/components/schemas/ParamSpec"
  /v1/pipelines/validate:
    post:
      summary: Validate a pipeline DAG
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PipelineDocumentV1"
      responses:
        "200":
          description: Validation result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ValidationResponse"
  /v1/pipelines/run:
    post:
      summary: Execute a pipeline
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RunRequest"
      responses:
        "200":
          description: Run result with artifacts
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RunResponse"
  /v1/projects/import/grip:
    post:
      summary: Import a legacy GRIP project
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        "200":
          description: GRIP import result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PipelineDocumentV1"
  /v1/codegen:
    post:
      summary: Generate code from a pipeline
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CodegenRequest"
      responses:
        "200":
          description: Code generation output
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: string
                  language:
                    type: string
  /v1/agent/generate:
    post:
      summary: AI-assisted pipeline generation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GenerateRequest"
      responses:
        "200":
          description: Agent generated pipeline
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GenerateResponse"
  /v1/sweeps:
    post:
      summary: Run a parameter sweep
      responses:
        "200":
          description: Sweep results
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    parameters:
                      type: object
                    score:
                      type: number
                    outputs:
                      type: object
  /v1/preview/events:
    post:
      responses:
        "200":
          description: Preview event stream payload
components:
  schemas:
    SocketSpec:
      type: object
      required: [name, type]
      properties:
        name:
          type: string
        type:
          type: string
          enum: [image, mask, number, boolean, contours, lines, json]
    ParamSpec:
      type: object
      required: [name, view, value]
      properties:
        name:
          type: string
        view:
          type: string
          enum: [text, slider, range, select, checkbox]
        value: {}
        options:
          type: array
          items:
            type: string
        min:
          type: number
        max:
          type: number
    EndpointRef:
      type: object
      required: [nodeId, socket]
      properties:
        nodeId:
          type: string
        socket:
          type: string
    PipelineNode:
      type: object
      required: [id, op]
      properties:
        id:
          type: string
        op:
          type: string
        label:
          type: string
        params:
          type: array
          items:
            $ref: "#/components/schemas/ParamSpec"
        inputs:
          type: array
          items:
            $ref: "#/components/schemas/SocketSpec"
        outputs:
          type: array
          items:
            $ref: "#/components/schemas/SocketSpec"
        position:
          type: object
          properties:
            x:
              type: number
            y:
              type: number
    PipelineEdge:
      type: object
      required: [id, from, to]
      properties:
        id:
          type: string
        from:
          $ref: "#/components/schemas/EndpointRef"
        to:
          $ref: "#/components/schemas/EndpointRef"
    PipelineDocumentV1:
      type: object
      required: [nodes, edges]
      properties:
        version:
          type: integer
          default: 1
        nodes:
          type: array
          items:
            $ref: "#/components/schemas/PipelineNode"
        edges:
          type: array
          items:
            $ref: "#/components/schemas/PipelineEdge"
        metadata:
          type: object
    ValidationResponse:
      type: object
      required: [ok]
      properties:
        ok:
          type: boolean
        errors:
          type: array
          items:
            type: string
    RunTrace:
      type: object
      required: [nodeId, status]
      properties:
        nodeId:
          type: string
        status:
          type: string
        message:
          type: string
    ArtifactRef:
      type: object
      required: [nodeId, socket, kind]
      properties:
        nodeId:
          type: string
        socket:
          type: string
        kind:
          type: string
        path:
          type: string
        value: {}
    RunRequest:
      type: object
      required: [pipeline]
      properties:
        pipeline:
          $ref: "#/components/schemas/PipelineDocumentV1"
        inputImagePath:
          type: string
    RunResponse:
      type: object
      required: [ok]
      properties:
        ok:
          type: boolean
        traces:
          type: array
          items:
            $ref: "#/components/schemas/RunTrace"
        outputs:
          type: object
        artifacts:
          type: array
          items:
            $ref: "#/components/schemas/ArtifactRef"
    CodegenRequest:
      type: object
      required: [pipeline, language]
      properties:
        pipeline:
          $ref: "#/components/schemas/PipelineDocumentV1"
        language:
          type: string
        className:
          type: string
          default: GeneratedPipeline
    GenerateRequest:
      type: object
      required: [requirements]
      properties:
        mode:
          type: string
          default: disabled
        requirements:
          type: string
        imageBase64:
          type: string
    GenerateResponse:
      type: object
      required: [pipeline]
      properties:
        pipeline:
          $ref: "#/components/schemas/PipelineDocumentV1"
        warnings:
          type: array
          items:
            type: string